1721: Refactor Admin Channel
