+++ packages/components/src/local/organisms/setting-channels/styles.module.scss
@@ -67,6 +67,12 @@
   }
 }
 
+.channelName {
+  input {
+    border-bottom: 1px solid;
+  }
+}
+
 .constraints-icon {
   width: 60px;
   text-align: center;
