+++ packages/components/src/local/organisms/setting-channels/types/props.d.ts
@@ -1,4 +1,4 @@
-import { ChannelTypes, ForceData, TemplateBody, ChannelCore } from '@serge/custom-types'
+import { ChannelTypes, ForceData, TemplateBody, ChannelCore, ChannelChat, ChannelCustom, ChannelMapping, ChannelCollab } from '@serge/custom-types'
 import { Option } from 'src/local/molecules/editable-row'
 
 export default interface PropTypes {
@@ -22,6 +22,34 @@ export type MessageGroupProps = {
   onDelete?: (value: string[]) => void
   type: string
   value: string[]
+}
+
+export type CollabChannelProps = {
+  channel: ChannelCollab
+  forces: ForceData[]
+  messageTemplates: TemplateBody[]
+  onChange: (channel: ChannelCollab) => void
+}
+
+export type ChatChannelProps = {
+  channel: ChannelChat
+  forces: ForceData[]
+  messageTemplates: TemplateBody[]
+  onChange: (channel: ChannelChat) => void
+}
+
+export type CustomChannelProps = {
+  channel: ChannelCustom
+  forces: ForceData[]
+  messageTemplates: TemplateBody[]
+  onChange: (channel: ChannelCustom) => void
+}
+
+export type MappingChannelProps = {
+  channel: ChannelMapping
+  forces: ForceData[]
+  messageTemplates: TemplateBody[]
+  onChange: (channel: ChannelMapping) => void
 }
 
 export { ChannelTypes, ChannelCollab, Channel, ForceData, Role, Participant } from '@serge/custom-types'
